backend/requirements.txt
langgraph==0.3.3
langchain==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.5


# Music API integrations
spotipy==2.22.1
tavily-python==0.3.3
requests==2.31.0

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data processing and ML
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
gensim==4.3.2
scipy==1.10.1


# Utilities
python-dotenv==1.0.0
pydantic>=2.7.4
